# docs/swagger.yaml
basePath: /
definitions:
...
  httpresponse.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  httpresponse.ErrorResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/httpresponse.ErrorDetail'
        type: array
      message:
        example: request validation failed
        type: string
    type: object
...
  student.updateRequest:
    type: object
  teacher.createRequest:
    properties:
      department:
        minLength: 3
        type: string
    required:
    - department
    type: object
  teacher.getResponse:
    properties:
      teacher:
        $ref: '#/definitions/teacher.teacherResponse'
    type: object
  teacher.listResponse:
    properties:
      teachers:
        items:
          $ref: '#/definitions/teacher.teacherResponse'
        type: array
    type: object
  teacher.teacherResponse:
    properties:
      department:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      userId:
        type: integer
    type: object
...
  user.updateRequest:
    properties:
      email:
        type: string
      firstName:
        maxLength: 20
        minLength: 1
        type: string
      lastName:
        maxLength: 20
        minLength: 1
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
  user.updateResponse:
    properties:
      user:
        $ref: '#/definitions/user.userResponse'
    type: object
  user.userResponse:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
    type: object
host: localhost:3800
info:
  contact:
    email: vanap387@gmail.com
    name: Ivan Popov
  description: API for managing courses, students, and teachers.
  termsOfService: http://swagger.io/terms/
  title: Study Platform API
  version: 1.0.0
paths:
...
  /students/me:
    put:
      consumes:
      - application/json
      description: Updates the authenticated student. The request schema is temporary
        and may change!!!
      parameters:
      - description: Temporary update current student request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/student.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "404":
          description: Student not found
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current student
      tags:
      - students
  /students/me/courses/{courseId}:
    delete:
      description: Removes the authenticated student from a course.
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "409":
          description: Student is not enrolled in this course
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unenroll from course
      tags:
      - students
    post:
      description: Enrolls the authenticated student in a course.
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "409":
          description: Student is already enrolled in this course
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enroll in course
      tags:
      - students
  /teachers:
    get:
      description: Returns the list of all teachers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teacher.listResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      summary: List teachers
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Creates a teacher profile for the authenticated user.
      parameters:
      - description: Create teacher request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/teacher.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "409":
          description: Teacher already exists
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create teacher profile
      tags:
      - teachers
  /teachers/{id}:
    get:
      description: Returns a teacher by user identifier.
      parameters:
      - description: Teacher user ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teacher.getResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "404":
          description: Teacher not found
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      summary: Get teacher
      tags:
      - teachers
...
  /users/me:
    delete:
      description: Deletes the authenticated user account.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile data and returns the updated
        user.
      parameters:
      - description: Update current user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.updateResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "409":
          description: User with this email already exists
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpresponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"